+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/__init__.py
@@ -44,6 +44,9 @@ from ._models import (
     CurrencyValue,
     CustomDocumentModelsDetails,
     ModelBuildMode,
+    ClassifierDocumentTypeDetails,
+    BlobSource,
+    BlobFileListSource,
     DocumentClassifierDetails,
     DocumentField,
     DocumentKeyValuePair,
@@ -69,14 +72,7 @@ from ._models import (
     TargetAuthorization,
     QuotaDetails,
     DocumentFormula,
-    DocumentImage,
     DocumentBarcode,
-    DocumentAnnotation,
-)
-from ._generated.models import (  # patched models
-    ClassifierDocumentTypeDetails,
-    AzureBlobFileListSource,
-    AzureBlobContentSource,
 )
 from ._api_versions import FormRecognizerApiVersion, DocumentAnalysisApiVersion
 
@@ -147,13 +143,11 @@ __all__ = [
     "DocumentAnalysisInnerError",
     "TargetAuthorization",
     "ClassifierDocumentTypeDetails",
-    "AzureBlobFileListSource",
-    "AzureBlobContentSource",
+    "BlobFileListSource",
+    "BlobSource",
     "QuotaDetails",
     "DocumentFormula",
-    "DocumentImage",
     "DocumentBarcode",
-    "DocumentAnnotation",
 ]
 
 __VERSION__ = VERSION