+++ sdk/ml/azure-ai-ml/azure/ai/ml/_arm_deployments/arm_helper.py
@@ -23,6 +23,7 @@ template_mapping = {
     ArmConstants.WORKSPACE_PARAM: "workspace_param.json",
     ArmConstants.FEATURE_STORE_ROLE_ASSIGNMENTS: "feature_store_role_assignments.json",
     ArmConstants.FEATURE_STORE_ROLE_ASSIGNMENTS_PARAM: "feature_store_role_assignments_param.json",
+    ArmConstants.WORKSPACE_PROJECT: "workspace_project.json",
 }
 
 
@@ -46,7 +47,7 @@ deployment_message_mapping = {
     ArmConstants.STORAGE_ACCOUNT_PARAMETER_NAME: "Creating Storage Account: ({0})",
     AzureMLResourceType.WORKSPACE: "Creating AzureML Workspace: ({0})",
     AzureMLResourceType.CONNECTIONS: "Creating connection: ({0})",
-    ArmConstants.USER_ASSIGNED_IDENTITES: "Creating User Assigned Identities: ({0})",
+    ArmConstants.USER_ASSIGNED_IDENTITIES: "Creating User Assigned Identities: ({0})",
 }