+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_workspace_operations_base.py
@@ -12,8 +12,8 @@ from azure.core.polling import LROPoller, PollingMethod
 
 from azure.ai.ml._arm_deployments import ArmDeploymentExecutor
 from azure.ai.ml._arm_deployments.arm_helper import get_template
-from azure.ai.ml._restclient.v2023_04_01_preview import AzureMachineLearningWorkspaces as ServiceClient042023Preview
-from azure.ai.ml._restclient.v2023_04_01_preview.models import (
+from azure.ai.ml._restclient.v2023_06_01_preview import AzureMachineLearningWorkspaces as ServiceClient062023Preview
+from azure.ai.ml._restclient.v2023_06_01_preview.models import (
     EncryptionKeyVaultUpdateProperties,
     EncryptionUpdateProperties,
     WorkspaceUpdateParameters,
@@ -38,6 +38,7 @@ from azure.ai.ml.constants._workspace import IsolationMode
 from azure.ai.ml.entities import Workspace
 from azure.ai.ml.entities._credentials import IdentityConfiguration
 from azure.ai.ml.entities._workspace.networking import ManagedNetwork
+from azure.ai.ml.entities._workspace_hub._constants import PROJECT_WORKSPACE_KIND, WORKSPACE_HUB_KIND
 from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationException
 
 ops_logger = OpsLogger(__name__)
@@ -50,7 +51,7 @@ class WorkspaceOperationsBase:
     def __init__(
         self,
         operation_scope: OperationScope,
-        service_client: ServiceClient042023Preview,
+        service_client: ServiceClient062023Preview,
         all_operations: OperationsContainer,
         credentials: Optional[TokenCredential] = None,
         **kwargs: Dict,
@@ -113,6 +114,26 @@ class WorkspaceOperationsBase:
 
         workspace.resource_group = resource_group
         template, param, resources_being_deployed = self._populate_arm_paramaters(workspace, **kwargs)
+        # check if create with workspace hub request is valid
+        if workspace._kind == PROJECT_WORKSPACE_KIND:
+            if not all(
+                x is None
+                for x in [
+                    workspace.storage_account,
+                    workspace.container_registry,
+                    workspace.key_vault,
+                    workspace.public_network_access,
+                    workspace.managed_network,
+                    workspace.customer_managed_key,
+                ]
+            ):
+                msg = "To create a project workspace with a workspace hub, please only specify name, description, display_name, location, application insight and identity"
+                raise ValidationException(
+                    message=msg,
+                    target=ErrorTarget.WORKSPACE,
+                    no_personal_data_message=msg,
+                    error_category=ErrorCategory.USER_ERROR,
+                )
 
         arm_submit = ArmDeploymentExecutor(
             credentials=self._credentials,
@@ -257,7 +278,7 @@ class WorkspaceOperationsBase:
                 and grant_materialization_identity_permissions
             ):
                 module_logger.info("updating feature store materialization identity role assignments..")
-                template, param, resources_being_deployed = self._populate_feature_store_role_assignemnt_parameters(
+                template, param, resources_being_deployed = self._populate_feature_store_role_assignment_parameters(
                     workspace, resource_group=resource_group, **kwargs
                 )
 
@@ -291,7 +312,16 @@ class WorkspaceOperationsBase:
     ) -> LROPoller[None]:
         workspace = self.get(name, **kwargs)
         resource_group = kwargs.get("resource_group") or self._resource_group_name
-        if delete_dependent_resources:
+
+        # prevent dependent resource delete for lean workspace, only delete appinsight
+        if workspace._kind == PROJECT_WORKSPACE_KIND and delete_dependent_resources:
+            delete_resource_by_arm_id(
+                self._credentials,
+                self._subscription_id,
+                workspace.application_insights,
+                ArmConstants.AZURE_MGMT_APPINSIGHT_API_VERSION,
+            )
+        elif delete_dependent_resources:
             delete_resource_by_arm_id(
                 self._credentials,
                 self._subscription_id,
@@ -316,6 +346,7 @@ class WorkspaceOperationsBase:
                 workspace.container_registry,
                 ArmConstants.AZURE_MGMT_CONTAINER_REG_API_VERSION,
             )
+
         poller = self._operation.begin_delete(
             resource_group_name=resource_group,
             workspace_name=name,
@@ -332,9 +363,10 @@ class WorkspaceOperationsBase:
             workspace.location = get_resource_group_location(
                 self._credentials, self._subscription_id, workspace.resource_group
             )
-
         template = get_template(resource_type=ArmConstants.WORKSPACE_BASE)
         param = get_template(resource_type=ArmConstants.WORKSPACE_PARAM)
+        if workspace._kind == PROJECT_WORKSPACE_KIND:
+            template = get_template(resource_type=ArmConstants.WORKSPACE_PROJECT)
         _set_val(param["workspaceName"], workspace.name)
         if not workspace.display_name:
             _set_val(param["friendlyName"], workspace.name)
@@ -485,7 +517,10 @@ class WorkspaceOperationsBase:
             if materialization_identity:
                 _set_val(param["materialization_identity_resource_id"], materialization_identity.resource_id)
             else:
-                _set_val(param["materializationIdentityName"], f"fstoreuai-{workspace.name}")
+                _set_val(
+                    param["materializationIdentityName"],
+                    f"materialization-uai-{workspace.resource_group}-{workspace.name}",
+                )
 
             if kwargs.get("grant_materialization_identity_permissions", None) is True:
                 _set_val(param["grant_materialization_identity_permissions"], "true")
@@ -499,10 +534,22 @@ class WorkspaceOperationsBase:
         if workspace.enable_data_isolation:
             _set_val(param["enable_data_isolation"], "true")
 
+        # Hub related param
+        if workspace._kind and workspace._kind.lower() == WORKSPACE_HUB_KIND:
+            if workspace.workspace_hub_config:
+                _set_val(param["workspace_hub_config"], workspace.workspace_hub_config._to_rest_object())
+            if workspace.existing_workspaces:
+                _set_val(param["existing_workspaces"], workspace.existing_workspaces)
+
+        # Lean related param
+        if workspace._kind and workspace._kind.lower() == PROJECT_WORKSPACE_KIND:
+            if workspace.workspace_hub:
+                _set_val(param["workspace_hub"], workspace.workspace_hub)
+
         resources_being_deployed[workspace.name] = (ArmConstants.WORKSPACE, None)
         return template, param, resources_being_deployed
 
-    def _populate_feature_store_role_assignemnt_parameters(
+    def _populate_feature_store_role_assignment_parameters(
         self, workspace: Workspace, **kwargs: Dict
     ) -> Tuple[dict, dict, dict]:
         resources_being_deployed = {}
@@ -537,7 +584,7 @@ class WorkspaceOperationsBase:
         if online_store_target:
             _set_val(param["online_store_target"], online_store_target)
 
-        resources_being_deployed[materialization_identity_id] = (ArmConstants.USER_ASSIGNED_IDENTITES, None)
+        resources_being_deployed[materialization_identity_id] = (ArmConstants.USER_ASSIGNED_IDENTITIES, None)
         return template, param, resources_being_deployed
 
     def _check_workspace_name(self, name) -> str: