+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -4,8 +4,12 @@
 ## 1.10.0 (Unreleased)
 
 ### Features Added
+- Added support of features that are known into the future/at forecast time for dnn in AutoML Forecasting jobs.
 
 ### Bugs Fixed
+- Local job runs will no longer fail if Docker registry has no username/password
+
+- Fixed an issue that code asset doesn't work with relative symbol links.
 
 ### Breaking Changes